Supply page children API endpoint
Also refactor cms API module structure, merge single_page and page modules, because they are too tightly coupled
(sharing the transform page and the get_single_page method)
